requirements.txt
certifi==2018.4.16
chardet==3.0.4
idna==2.6
multi-key-dict==2.0.3
pbr==4.0.2
python-jenkins==1.0.0
requests==2.18.4
six==1.11.0
urllib3==1.22
